+++ source/Public/Test-SqlDscIsLoginEnabled.ps1
@@ -1,0 +1,86 @@
+<#
+    .SYNOPSIS
+        Returns whether the server login is enabled or disabled.
+
+    .DESCRIPTION
+        Returns whether the server login is enabled or disabled.
+
+    .PARAMETER ServerObject
+        Specifies current server connection object.
+
+    .PARAMETER LoginObject
+        Specifies a login object to test.
+
+    .PARAMETER Name
+        Specifies the name of the server login to test.
+
+    .PARAMETER Refresh
+        Specifies that the **ServerObject**'s logins should be refreshed before
+        trying to test the login object. This is helpful when logins could have
+        been modified outside of the **ServerObject**, for example through T-SQL.
+        But on instances with a large amount of logins it might be better to make
+        sure the **ServerObject** is recent enough, or pass in **LoginObject**.
+
+    .OUTPUTS
+        [System.Boolean]
+
+    .EXAMPLE
+        $serverObject = Connect-SqlDscDatabaseEngine -InstanceName 'MyInstance'
+        Test-SqlDscIsLoginEnabled -ServerObject $serverObject -Name 'MyLogin'
+
+        Returns $true if the login is enabled, if not $false is returned.
+
+    .EXAMPLE
+        $serverObject = Connect-SqlDscDatabaseEngine -InstanceName 'MyInstance'
+        $loginObject = $serverObject | Get-SqlDscLogin -Name 'MyLogin'
+        Test-SqlDscIsLoginEnabled -LoginObject $loginObject
+
+        Returns $true if the login is enabled, if not $false is returned.
+#>
+function Test-SqlDscIsLoginEnabled
+{
+    [System.Diagnostics.CodeAnalysis.SuppressMessageAttribute('UseSyntacticallyCorrectExamples', '', Justification = 'Because the rule does not yet support parsing the code when a parameter type is not available. The ScriptAnalyzer rule UseSyntacticallyCorrectExamples will always error in the editor due to https://github.com/indented-automation/Indented.ScriptAnalyzerRules/issues/8.')]
+    [CmdletBinding()]
+    [OutputType([System.Boolean])]
+    param
+    (
+        [Parameter(ParameterSetName = 'ServerObject', Mandatory = $true, ValueFromPipeline = $true)]
+        [Microsoft.SqlServer.Management.Smo.Server]
+        $ServerObject,
+
+        [Parameter(ParameterSetName = 'LoginObject', Mandatory = $true, ValueFromPipeline = $true)]
+        [Microsoft.SqlServer.Management.Smo.Login]
+        $LoginObject,
+
+        [Parameter(ParameterSetName = 'ServerObject', Mandatory = $true)]
+        [System.String]
+        $Name,
+
+        [Parameter(ParameterSetName = 'ServerObject')]
+        [System.Management.Automation.SwitchParameter]
+        $Refresh
+    )
+
+    process
+    {
+        if ($PSCmdlet.ParameterSetName -eq 'ServerObject')
+        {
+            $getSqlDscLoginParameters = @{
+                ServerObject = $ServerObject
+                Name         = $Name
+                Refresh      = $Refresh
+                ErrorAction  = 'Stop'
+            }
+
+            # If this command does not find the login it will throw an exception.
+            $loginObjectArray = Get-SqlDscLogin @getSqlDscLoginParameters
+
+            # Pick the only object in the array.
+            $LoginObject = $loginObjectArray | Select-Object -First 1
+        }
+
+        $loginEnabled = -not $LoginObject.IsDisabled
+
+        return $loginEnabled
+    }
+}